Change MaxPostSize to int64
Summary:
When people try to increase the limit from 1GB to 2GB, the number becomes
negative and cause confusing error messages such as "POST Content-Length of
83128383 bytes exceeds the limit of 3695181824 bytes"

Task ID: #

Blame Rev:

Reviewers:
hzhao
CC:
hphp-diffs@lists.facebook.com
Test Plan:
make fast_tests
no more that confusing error message
upload a video successfully under normal mode

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 195988

+++ src/runtime/base/runtime_option.cpp
@@ -112,7 +112,7 @@ std::string RuntimeOption::OutputHandler;
 bool RuntimeOption::ImplicitFlush = false;
 bool RuntimeOption::EnableEarlyFlush = true;
 bool RuntimeOption::ForceChunkedEncoding = false;
-int RuntimeOption::MaxPostSize;
+int64 RuntimeOption::MaxPostSize;
 bool RuntimeOption::AlwaysPopulateRawPostData = true;
 int64 RuntimeOption::UploadMaxFileSize;
 std::string RuntimeOption::UploadTmpDir;
@@ -570,7 +570,7 @@ void RuntimeOption::Load(Hdf &config, StringVec *overwrites /* = NULL */) {
     ImplicitFlush = server["ImplicitFlush"].getBool();
     EnableEarlyFlush = server["EnableEarlyFlush"].getBool(true);
     ForceChunkedEncoding = server["ForceChunkedEncoding"].getBool();
-    MaxPostSize = (server["MaxPostSize"].getInt32(100)) * (1 << 20);
+    MaxPostSize = (server["MaxPostSize"].getInt32(100)) * (1LL << 20);
     AlwaysPopulateRawPostData = server["AlwaysPopulateRawPostData"].getBool();
     LibEventSyncSend = server["LibEventSyncSend"].getBool(true);
     TakeoverFilename = server["TakeoverFilename"].getString();
@@ -702,7 +702,7 @@ void RuntimeOption::Load(Hdf &config, StringVec *overwrites /* = NULL */) {
 
     Hdf upload = server["Upload"];
     UploadMaxFileSize =
-      (upload["UploadMaxFileSize"].getInt32(100)) * (1 << 20);
+      (upload["UploadMaxFileSize"].getInt32(100)) * (1LL << 20);
     UploadTmpDir = upload["UploadTmpDir"].getString("/tmp");
     RuntimeOption::AllowedDirectories.push_back(UploadTmpDir);
     EnableFileUploads = upload["EnableFileUploads"].getBool(true);
